+++ src/Api/Endpoints/Zaken/Zaakverzoeken.php
@@ -6,22 +6,19 @@ namespace Woweb\Zgw\Api\Endpoints\Zaken;
 
 use Woweb\Zgw\Api\Actions\Delete;
 use Woweb\Zgw\Api\Actions\Index;
-use Woweb\Zgw\Api\Actions\Patch;
-use Woweb\Zgw\Api\Actions\Put;
 use Woweb\Zgw\Api\Actions\Show;
 use Woweb\Zgw\Api\Actions\Store;
 use Woweb\Zgw\Api\Endpoints\AbstractEndpoint;
 
-class Zaakobjecten extends AbstractEndpoint
+// A relation between a zaak and a verzoek: create, read and delete.
+class Zaakverzoeken extends AbstractEndpoint
 {
     use Delete;
     use Index;
-    use Patch;
-    use Put;
     use Show;
     use Store;
 
     protected string $apiName = 'zaken';
 
-    protected string $endpoint = 'zaakobjecten';
+    protected string $endpoint = 'zaakverzoeken';
 }